implement CLayoutEngine::LayoutNormal() using DeferWindowPos(), to simplify layout of top-level dialogs;  remove unused layout::DoRepaint constant;  extract layout::CDiagnostics for debugging.

+++ CommonUtl/utl/FlagTags.h
@@ -14,7 +14,7 @@ class CFlagTags
 public:
 	struct FlagDef
 	{
-		int m_flag;
+		unsigned long m_flag;
 		const TCHAR* m_pKeyTag;
 	};
